// app/[locale]/dashboard/components/entries-masonry.skeleton-text.generated.ts
type GeneratedSkeletonTextEntry = {
  lines: GeneratedSkeletonTextLines;
  barWidth?: GeneratedSkeletonTextBarWidth;
  lastBarWidth?: GeneratedSkeletonTextLastBarWidth;
  barHeight?: GeneratedSkeletonTextBarHeight;
  lineHeight?: GeneratedSkeletonTextLineHeight;
  responsiveBy?: GeneratedSkeletonTextResponsiveBy;
};
type GeneratedSkeletonResponsiveNumber = number | Record<number, number>;
type GeneratedSkeletonTextModule = Record<string, GeneratedSkeletonTextEntry>;
export const entriesMasonryEntry01Section: GeneratedSkeletonTextEntry = {
  "lines": 1,
  "barWidth": "47px",
  "barHeight": 11.52,
  "lineHeight": 1.5
};
export const entriesMasonryEntry01Title: GeneratedSkeletonTextEntry = {
  "lines": {
    "0": 2,
    "392": 1
  },
  "barWidth": {
    "0": [
      "100px",
      "104px"
    ],
    "392": "208px"
  },
  "barHeight": 17.28,
  "lineHeight": 1.2
};
export const entriesMasonryEntry01Count: GeneratedSkeletonTextEntry = {
  "lines": 1,
  "barWidth": "53px",
  "barHeight": 12.48,
  "lineHeight": 1.5
};
export const entriesMasonryEntry01Body: GeneratedSkeletonTextEntry = {
  "lines": {
    "0": 4,
    "347": 3,
    "480": 2,
    "1088": 1
  },
  "barWidth": {
    "0": [
      "195px",
      "170px",
      "206px",
      "100px"
    ],
    "347": [
      "226px",
      "232px",
      "217px"
    ],
    "480": [
      "314px",
      "365px"
    ],
    "1088": "684px"
  },
  "barHeight": 15.2,
  "lineHeight": 1.65
};
export const entriesMasonryEntry02Section: GeneratedSkeletonTextEntry = {
  "lines": 1,
  "barWidth": "44px",
  "barHeight": 11.52,
  "lineHeight": 1.5
};
export const entriesMasonryEntry02Title: GeneratedSkeletonTextEntry = {
  "lines": {
    "0": 2,
    "377": 1
  },
  "barWidth": {
    "0": [
      "134px",
      "56px"
    ],
    "377": "193px"
  },
  "barHeight": 17.28,
  "lineHeight": 1.2
};
export const entriesMasonryEntry02Count: GeneratedSkeletonTextEntry = {
  "lines": 1,
  "barWidth": "53px",
  "barHeight": 12.48,
  "lineHeight": 1.5
};
export const entriesMasonryEntry02Body: GeneratedSkeletonTextEntry = {
  "lines": {
    "0": 4,
    "321": 3,
    "412": 2,
    "713": 1,
    "768": 2,
    "998": 1
  },
  "barWidth": {
    "0": [
      "179px",
      "155px",
      "166px",
      "80px"
    ],
    "321": [
      "179px",
      "207px",
      "199px"
    ],
    "412": [
      "292px",
      "297px"
    ],
    "713": "594px",
    "768": [
      "339px",
      "250px"
    ],
    "998": "594px"
  },
  "barHeight": 15.2,
  "lineHeight": 1.65
};
export const entriesMasonryEntry03Section: GeneratedSkeletonTextEntry = {
  "lines": 1,
  "barWidth": "47px",
  "barHeight": 11.52,
  "lineHeight": 1.5
};
export const entriesMasonryEntry03Title: GeneratedSkeletonTextEntry = {
  "lines": {
    "0": 2,
    "338": 1
  },
  "barWidth": {
    "0": [
      "111px",
      "39px"
    ],
    "338": "154px"
  },
  "barHeight": 17.28,
  "lineHeight": 1.2
};
export const entriesMasonryEntry03Count: GeneratedSkeletonTextEntry = {
  "lines": 1,
  "barWidth": "53px",
  "barHeight": 12.48,
  "lineHeight": 1.5
};
export const entriesMasonryEntry03Body: GeneratedSkeletonTextEntry = {
  "lines": {
    "0": 3,
    "389": 2,
    "656": 1,
    "768": 2,
    "942": 1
  },
  "barWidth": {
    "0": [
      "203px",
      "186px",
      "143px"
    ],
    "389": [
      "274px",
      "261px"
    ],
    "656": "540px",
    "768": [
      "371px",
      "165px"
    ],
    "942": "540px"
  },
  "barHeight": 15.2,
  "lineHeight": 1.65
};
export const entriesMasonrySkeletonText: GeneratedSkeletonTextModule = {
  entriesMasonryEntry01Section,
  entriesMasonryEntry01Title,
  entriesMasonryEntry01Count,
  entriesMasonryEntry01Body,
  entriesMasonryEntry02Section,
  entriesMasonryEntry02Title,
  entriesMasonryEntry02Count,
  entriesMasonryEntry02Body,
  entriesMasonryEntry03Section,
  entriesMasonryEntry03Title,
  entriesMasonryEntry03Count,
  entriesMasonryEntry03Body,
};